Ticket #7130: Log entries related to assignments

+++ app/models/concerns/assignment_concern.rb
@@ -11,16 +11,19 @@ module AssignmentConcern
   end
 
   # We can't simply use assignments= from Active Record here because it doesn't call the callbacks
-  def assigned_to_ids=(csids)
-    new_ids = csids.to_s.split(',').map(&:to_i)
-    current_ids = self.reload.assignments.map(&:user_id)
-    to_create = new_ids - current_ids
-    to_delete = current_ids - new_ids
-    to_delete.each do |id|
-      Assignment.where(annotation_id: self.id, user_id: id).last.destroy!
-    end
-    to_create.each do |id|
-      Assignment.create!(annotation_id: self.id, user_id: id)
+  def save_assignments
+    csids = self.assigned_to_ids
+    unless csids.nil?
+      new_ids = csids.to_s.split(',').map(&:to_i)
+      current_ids = self.reload.assignments.map(&:user_id)
+      to_create = new_ids - current_ids
+      to_delete = current_ids - new_ids
+      to_delete.each do |id|
+        Assignment.where(annotation_id: self.id, user_id: id).last.destroy!
+      end
+      to_create.each do |id|
+        Assignment.create!(annotation_id: self.id, user_id: id)
+      end
     end
   end
 
@@ -48,6 +51,10 @@ module AssignmentConcern
   end
 
   included do
+    attr_accessor :assigned_to_ids
+
+    after_save :save_assignments
+
     has_many :assignments, foreign_key: :annotation_id, dependent: :destroy
   end
 end
